delete GPSodom_v1 gps_common
