--- gps_common/launch/fix_translator.launch
@@ -1,30 +1,0 @@
-<launch>
-
-<!--
-    fix_translator translates to and from NatSatFix 
-    and GPSFix messages.
-
-    If you want to translate from NavSatFix to GPSFix,
-    you have to modify the first two remap lines.
-   
-    If you want to translate from GPSFix to NavSatFix,
-    you have to uncomment and modify the last two remap 
-    lines.
-
-    Only adjust topic names after "to=" in each remap line.
--->
-
-  <node name="fix_translator" pkg="gps_common" type="fix_translator">
-    <!-- Translate from NavSatFix to GPSFix //-->
-      <remap from="/navsat_fix_in"  to="/fix"/>       
-      <remap from="/gps_fix_out"    to="/gps_fix"/>
-
-    <!-- Translate from GPSFix to NavSatFix //-->
-     <!--
-       <remap from="/gps_fix_in"     to="/YOUR_GPSFIX_TOPIC"/>   
-       <remap from="/navsat_fix_out" to="/YOUR_NAVSATFIX_TOPIC"/>   
-     //-->
-  </node>
-
-</launch>
-
