--- gps_common/msg/GPSFix.msg
@@ -1,89 +1,0 @@
-# A more complete GPS fix to supplement sensor_msgs/NavSatFix.
-Header header
-
-GPSStatus status
-
-# Latitude (degrees). Positive is north of equator; negative is south.
-float64 latitude
-
-# Longitude (degrees). Positive is east of prime meridian, negative west.
-float64 longitude
-
-# Altitude (meters). Positive is above reference (e.g., sea level).
-float64 altitude
-
-# Direction (degrees from north)
-float64 track
-
-# Ground speed (meters/second)
-float64 speed
-
-# Vertical speed (meters/second)
-float64 climb
-
-# Device orientation (units in degrees)
-float64 pitch
-float64 roll
-float64 dip
-
-# GPS time
-float64 time
-
-## Dilution of precision; Xdop<=0 means the value is unknown
-
-# Total (positional-temporal) dilution of precision
-float64 gdop
-
-# Positional (3D) dilution of precision
-float64 pdop
-
-# Horizontal dilution of precision
-float64 hdop
-
-# Vertical dilution of precision
-float64 vdop
-
-# Temporal dilution of precision
-float64 tdop
-
-## Uncertainty of measurement, 95% confidence
-
-# Spherical position uncertainty (meters) [epe]
-float64 err
-
-# Horizontal position uncertainty (meters) [eph]
-float64 err_horz
-
-# Vertical position uncertainty (meters) [epv]
-float64 err_vert
-
-# Track uncertainty (degrees) [epd]
-float64 err_track
-
-# Ground speed uncertainty (meters/second) [eps]
-float64 err_speed
-
-# Vertical speed uncertainty (meters/second) [epc]
-float64 err_climb
-
-# Temporal uncertainty [ept]
-float64 err_time
-
-# Orientation uncertainty (degrees)
-float64 err_pitch
-float64 err_roll
-float64 err_dip
-
-# Position covariance [m^2] defined relative to a tangential plane
-# through the reported position. The components are East, North, and
-# Up (ENU), in row-major order.
-
-float64[9] position_covariance
-
-uint8 COVARIANCE_TYPE_UNKNOWN = 0
-uint8 COVARIANCE_TYPE_APPROXIMATED = 1
-uint8 COVARIANCE_TYPE_DIAGONAL_KNOWN = 2
-uint8 COVARIANCE_TYPE_KNOWN = 3
-
-uint8 position_covariance_type
-
