--- gps_common/package.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>gps_common</name>
-  <version>0.3.4</version>
-  <description>GPS messages and common routines for use in GPS drivers</description>
-
-  <maintainer email="timo.roehling@fkie.fraunhofer.de">Timo Roehling</maintainer>
-  <maintainer email="preed@swri.org">P. J. Reed</maintainer>
-  <license>BSD</license>
-  <url type="website">http://ros.org/wiki/gps_common</url>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-  <build_depend>message_filters</build_depend>
-  <build_depend>message_generation</build_depend>
-  <build_depend>nav_msgs</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>sensor_msgs</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>rospy</build_depend>
-
-  <run_depend>message_filters</run_depend>
-  <run_depend>message_runtime</run_depend>
-  <run_depend>nav_msgs</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>sensor_msgs</run_depend>
-  <run_depend>std_msgs</run_depend>
-  <run_depend>rospy</run_depend>
-
-</package>
