--- gps_common/src/gps_common/__init__.py
@@ -1,1 +1,0 @@
-# Intentionally empty
